Implement service worker caching
Assets are stored as a versioned URL, yet requested with an un-versioned URL. The service worker appends a version hash to the URL request, which is generated on build time.

This scheme requires that service worker support be available and that HTTPS is used. The WASM module is not loaded until a service worker is controlling requests for the page.

Later, configuration can be made to disable service worker support.